* NetBSD compatibility * Correct broken file open * Add owner,group and perm options for output files * type correctness for tolower conversions

+++ options.c
@@ -398,7 +398,7 @@ int add_common_opt(char *var, char *val,iface_t *ifp)
             return(-1);
         /* Normalize to lower case */
         for (ptr=ifp->name;*val;)
-                *ptr++=tolower(*val++);
+                *ptr++=tolower((int) *val++);
     } else
         return(1);
